feat: move drive to trash

+++ DigitalAssetManagement.Infrastructure/Services/DriveServiceImplementation.cs
@@ -6,6 +6,8 @@ using DigitalAssetManagement.Application.Exceptions;
 using DigitalAssetManagement.Application.Repositories;
 using DigitalAssetManagement.Application.Services;
 using DigitalAssetManagement.Domain.Entities;
+using Hangfire;
+using Microsoft.Extensions.Configuration;
 using System.Linq.Expressions;
 
 namespace DigitalAssetManagement.Infrastructure.Services
@@ -15,12 +17,16 @@ namespace DigitalAssetManagement.Infrastructure.Services
         private readonly UnitOfWork _unitOfWork;
         private readonly IMapper _mapper;
         private readonly UserService _userService;
+        private readonly IBackgroundJobClient _backgroundJobClient;
+        private readonly IConfiguration _configuration;
 
-        public DriveServiceImplementation(UnitOfWork unitOfWork, IMapper mapper, UserService userService)
+        public DriveServiceImplementation(UnitOfWork unitOfWork, IMapper mapper, UserService userService, IBackgroundJobClient backgroundJobClient, IConfiguration configuration)
         {
             _unitOfWork = unitOfWork;
             _mapper = mapper;
             _userService = userService;
+            _backgroundJobClient = backgroundJobClient;
+            _configuration = configuration;
         }
 
         public async Task<DriveDetailsResponseDto> Create(DriveRequestDto request)
@@ -51,6 +57,18 @@ namespace DigitalAssetManagement.Infrastructure.Services
                 throw new ForbiddenException(ExceptionMessage.UnallowedModification);
             }
 
+            _unitOfWork.DriveRepository.Delete(drive);
+            await _unitOfWork.SaveAsync();
+        }
+
+        public async Task DeleteDrive(int id)
+        {
+            var drive = await _unitOfWork.DriveRepository.GetByIdAsync(id);
+            if (drive == null)
+            {
+                throw new NotFoundException(ExceptionMessage.DriveNotFound);
+            }
+            
             _unitOfWork.DriveRepository.Delete(drive);
             await _unitOfWork.SaveAsync();
         }
@@ -99,6 +117,28 @@ namespace DigitalAssetManagement.Infrastructure.Services
 
             var drives = await _unitOfWork.DriveRepository.GetAllAsync(filter: Expression.Lambda<Func<Drive, bool>>(expression, parameter));
             return _mapper.Map<List<DriveResponseDto>>(drives);
+        }
+
+        public async Task MoveToTrash(int id)
+        {
+            var drive = await _unitOfWork.DriveRepository.GetByIdAsync(id);
+            if (drive == null)
+            {
+                throw new NotFoundException(ExceptionMessage.DriveNotFound);
+            }
+
+            var user = await _userService.GetLoginUserAsync();
+            if (drive.OwnerId != user.Id)
+            {
+                throw new ForbiddenException(ExceptionMessage.UnallowedModification);
+            }
+
+            // TODO: change status of subfolders & file
+            drive.IsDeleted = true;
+            _unitOfWork.DriveRepository.Update(drive);
+            await _unitOfWork.SaveAsync();
+
+            _backgroundJobClient.Schedule(() => this.DeleteDrive(id), TimeSpan.FromDays(int.Parse(_configuration["schedule:deletedWaitDays"]!)));
         }
     }
 }
